+++ src/rversion.rs
@@ -10,7 +10,9 @@ pub struct Rversion {
 pub struct LinuxVersion {
     pub distro: String,
     pub version: String,
-    pub url: String
+    pub url: String,
+    pub rspm: bool,
+    pub rspm_url: String
 }
 
 #[cfg(any(target_os = "macos", target_os = "linux"))]